-board定義変更 -Arduino 1.6.5->1.8.10変更

+++ libraries/esp32/robot_pcmode/robot_pcmode.ino
@@ -115,7 +115,7 @@ static void callOK()
 {
     Serial.write(0xff);
     Serial.write(0x55);
-    Serial.write(0);
+    Serial.write((uint8_t)0);
 }
 
 static void sendByte(uint8_t data)
